Improve handling of the tempo of a music expression.
Finally figured out how to handle time signatures: For example, with x/y, where x is >= 6 and a multiple of 3 (6, 9, 12, ...), then the time signature is compound. Otherwise it is simple.

In simple time, the y indicates that each beat corresponds to a note of length 1/y. In compound time, the y indicates that each beat corresponds to a note of length 3/y.

Next I might study the possibility of using Helmholtz pitch notation (C,, C, C c c′ c''c''') to indicate octaves, and use ABC notation to indicate length (similar to what is used right now, but more advanced).