Added chords. Strange timing issues when playing were solved.
This was accomplished by reducing the value of  AudioPeriodSize (buffer size) from 4096 to 1024 (512 seemed to work too, but was more prone to stutters if the CPU was under heavier load). Also, the thread sleeping method was changed to a more high-precision one (mixing the OS sleep method for a large portion of the waiting time, with a SpinLock-like loop for the last milliseconds before executing a note). A more in depth explanation of the change is included in the comments of the MidiPlayer.