Initial implementation of variables
Expanded the grammar to include statements (instead of just music expressions, like before). Each statement must be separated by ; to keep the grammar simple, at least for now.

Currently there are two statements implemented: variable declaration and instrument declaration.

Although instrument declaration already works, the program still needs to be manually set in the source code, because the default fluidsynth sequencer does not allow setting sendind an event to change the program mid-sequence. In a future commit, this will be fixed by implementing a custom-made sequencer that will run on a seperate thread.

The variable declaration should already work. For now the only values that variables support are musical expressions, since there are no other values implemented right now.

Getting variables to work was a little bit tricky, and it's implementation is not yet final. The issue is that since musical expressions are lazy, they are only generated on the fly. But this means that if the state changes between a variable declaration and it's use somewhere else, the notes would reflect those state changes. In essence, variables need a frozen state (which is partly implemented for the context class, but not for the symbols scope).