Add functions and value expressions.
Right now there are two kinds of expressions (value and musical). In some places they can both be used, while others require a specific kind of expression. This was done because musicas expressions have very specific semantics for the same operators, which means that keeping them separated was needed for now. This can change in the future however.

Right now functions can receive parameters and return music or values. The parser is still a little rough around the edges, especially because a lot of invalid syntax still get's correctly parsed (reason currently unknown). This is especially apparent when one forgets a `;` (either after a statement or between arguments)

This leads to another issue we are facing: most languages use commas `,` to separate arguments when calling functions, but since commas are already taken for our musical notes syntax, `;` were used. Not sure if this can be confusing thoug, needs more thinking.

Also, musical expressions are played by default. However, some expressions can be ambiguous (contain musical **values** but not be necessarily a musical expression, like variables for example). For this reason, there is the `play` function which evaluates and plays the notes of the expression, regardless of what type of AST node it is. Other conveniency functions already included are the `discard` and `debug` functions (altough the debugging of notes still needs some working on).

With this foundation I can explore the construction of virtual keyboards within the language, as well as evaluate if there is the need for syntatic sugar for them or if calling functions is enough.

Also will start to explore other forms of interface (terminal) and different outputs (to music files, MIDI, music sheets, etc...) for the language.