Improve music events and player sequencer
Refactored a lot of code to make it easier to export to different formats (by emitting more events, which contain useful metadata when exporting to ABC, for instance) and by moving the fluidsynth's sequencer behind a sequencer abstraction of my own. This will allow me to create sequencers that don't generate sound, but for instance a MIDI or a text file. Also the player now supports having more than one sequencer (which in the near future will mean the cli application will allow setting more than one output at the same time).